+++ MapboxNavigation/MBRouteVoiceController.m
@@ -1,3 +1,0 @@
-#import "MBRouteVoiceController.h"
-
-const NSErrorUserInfoKey MBSpokenInstructionErrorCodeKey = @"MBSpokenInstructionErrorCode";