+++ MapboxNavigation/MapboxNavigation.h
@@ -6,5 +6,5 @@ FOUNDATION_EXPORT double MapboxNavigationVersionNumber;
 //! Project version string for MapboxNavigation.
 FOUNDATION_EXPORT const unsigned char MapboxNavigationVersionString[];
 
-#import "MBRouteVoiceController.h"
+#import "MBConstants.h"
 #import "MGLMapView+MGLNavigationAdditions.h"